nwbの保存を変更

+++ optinist/wrappers/suite2p_wrapper/registration.py
@@ -19,21 +19,19 @@ def suite2p_registration(
         refImg = refImg[0]
 
     ops = {**default_ops(), **ops, **params}
-    # import pdb; pdb.set_trace()
 
     # register binary
     ops = registration.register_binary(ops, refImg=refImg)
 
     # compute metrics for registration
-    # import pdb; pdb.set_trace()
     if ops.get('do_regmetrics', True) and ops['nframes']>=1500:
         ops = registration.get_pc_metrics(ops)
 
     info = {}
     images = io.BinaryFile(read_filename=ops['reg_file'], Ly=ops['Ly'], Lx=ops['Lx']).data
     info['images'] = ImageData(images, func_name='suite2p_registration', file_name='images')
-    info['refImg'] = ImageData(ops['refImg'].astype(np.uint8), func_name='suite2p_registration', file_name='refImg')
-    info['meanImgE'] = ImageData(ops['meanImgE'].astype(np.uint8), func_name='suite2p_registration', file_name='meanImgE')
+    info['refImg'] = ImageData(ops['refImg'], func_name='suite2p_registration', file_name='refImg')
+    info['meanImgE'] = ImageData(ops['meanImgE'], func_name='suite2p_registration', file_name='meanImgE')
     info['ops'] = Suite2pData(ops, func_name='suite2p_registration', file_name='ops')
 
     return info
